Debugging database object lookups/database updates.

+++ js_builtins.c
@@ -102,7 +102,7 @@ value_t eval_builtin(Node *a, environment_t *env) {
 
 	idx = fc->hdr->aux;
 	name = builtIns[idx].name;
-	fn = findFirstNode(env->table, a - env->table);
+	fn = findFirstNode(env->table, (uint32_t)(a - env->table));
 
 	fprintf (stderr, "File: %s, Line: %d, Function: %s Status: %s\n", fn->script, (int)a->lineNo, name, strstatus(v.status));
 	exit(1);
